[new-hotspot-scheduler] add operator influence and more metrics (#1713)
* *: unify get store function everywhere (#1671)


* remove unnecessary parentheses

*  server: use leader lease to determine tso service validity (#1676)


* change internal stat values to float64

* add pending operator influence

* add metrics of pending influence

* fix metrics

* fix panic

* adjust pending influence of balanceHotWrite

* change weight of pending influence

* test: fix tests (#1696)

* test: fix region syncer test


* decrease region rolling window; store pending influence in scheduler

* add config-check flag for pd-server (#1695)


* decrease possiblility transfer hot write leader

* change pending influence weight

* add unstarted op metrics

* add logs for debug

* add log for debug

* add logs for debug

* add logs for debug

* add logs for debug

* add logs for debug

* add logs for debug

* add logs for debug

* Revert "add logs for debug"

This reverts commit e74c7a9a2a4d8aacd0a85fea65d1a33e7de10e19.

* add metrics for hotspot operators

* operator: rewrite move region related functions (#1667)

* add metrics for pending operators

* *: support setting endKey for ScanRange (#1700)


* fix bug

* fix bug

* fix bug

* fix metrics thread-safe bug

* fix logic bug

* *: reduce some unnecessary parameters (#1698)


* schedule: Do not send an operator of a region wth a stale epoch (#1659)

* schedule: Do not send an operator of a region wth a stale epoch


* schedule: check the version changed by the operator self


* schedule: fix unit test


* schedule: fix to avoid dispatching a stale opstep


* dispatch: refactor "ConsumeConfVer() int" to "ExpectConfVerChange() bool"


* dispatch: fix typo in comment


* fix typo


* dispatch: fix unittest


* dispatch: refine format


* server: fix the dead lock in scatter region (#1706)


* add drop time for operator

* use IsDropped to recognize canceled ops

* try to fix trans leader burst

* try to fix trans leader burst

* add zombie influence

* change select src dst strategy; improve op_controller

* change select src strategy

* fix bug

* tools: fix set namespace in pd-ctl (#1701)


* tools: fix parse url without http prefix (#1703)


* tests: support deadlock detection in make test (#1704)


* Makefile: fix failpoint enable (#1722)


* checker: fix the issue that a region does not merge to the sibling with smaller size (#1723)


* tools: balance region simulator (#1708)

* scheduler: do not remove the operator when the step does not finish (#1715)


* operator: fix the AddLearner config version judgment (#1732)


* tools: fix TLS in pd control (#1729)


* syncer: support TLS for region syncer (#1728)


* schedule: fix a thread-safe bug and improve code (#1719)

+++ server/statistics/store.go
@@ -133,80 +133,80 @@ func (s *StoresStats) GetStoreDiskWriteRate(storeID uint64) float64 {
 }
 
 // GetStoresCPUUsage returns the cpu usage stat of all StoreInfo.
-func (s *StoresStats) GetStoresCPUUsage() map[uint64]uint64 {
+func (s *StoresStats) GetStoresCPUUsage() map[uint64]float64 {
 	s.RLock()
 	defer s.RUnlock()
-	res := make(map[uint64]uint64, len(s.rollingStoresStats))
+	res := make(map[uint64]float64, len(s.rollingStoresStats))
 	for storeID, stats := range s.rollingStoresStats {
-		res[storeID] = uint64(stats.GetCPUUsage())
+		res[storeID] = stats.GetCPUUsage()
 	}
 	return res
 }
 
 // GetStoresDiskReadRate returns the disk read rate stat of all StoreInfo.
-func (s *StoresStats) GetStoresDiskReadRate() map[uint64]uint64 {
+func (s *StoresStats) GetStoresDiskReadRate() map[uint64]float64 {
 	s.RLock()
 	defer s.RUnlock()
-	res := make(map[uint64]uint64, len(s.rollingStoresStats))
+	res := make(map[uint64]float64, len(s.rollingStoresStats))
 	for storeID, stats := range s.rollingStoresStats {
-		res[storeID] = uint64(stats.GetDiskReadRate())
+		res[storeID] = stats.GetDiskReadRate()
 	}
 	return res
 }
 
 // GetStoresDiskWriteRate returns the disk write rate stat of all StoreInfo.
-func (s *StoresStats) GetStoresDiskWriteRate() map[uint64]uint64 {
+func (s *StoresStats) GetStoresDiskWriteRate() map[uint64]float64 {
 	s.RLock()
 	defer s.RUnlock()
-	res := make(map[uint64]uint64, len(s.rollingStoresStats))
+	res := make(map[uint64]float64, len(s.rollingStoresStats))
 	for storeID, stats := range s.rollingStoresStats {
-		res[storeID] = uint64(stats.GetDiskWriteRate())
+		res[storeID] = stats.GetDiskWriteRate()
 	}
 	return res
 }
 
 // GetStoresBytesWriteStat returns the bytes write stat of all StoreInfo.
-func (s *StoresStats) GetStoresBytesWriteStat() map[uint64]uint64 {
+func (s *StoresStats) GetStoresBytesWriteStat() map[uint64]float64 {
 	s.RLock()
 	defer s.RUnlock()
-	res := make(map[uint64]uint64, len(s.rollingStoresStats))
+	res := make(map[uint64]float64, len(s.rollingStoresStats))
 	for storeID, stats := range s.rollingStoresStats {
 		writeRate, _ := stats.GetBytesRate()
-		res[storeID] = uint64(writeRate)
+		res[storeID] = writeRate
 	}
 	return res
 }
 
 // GetStoresBytesReadStat returns the bytes read stat of all StoreInfo.
-func (s *StoresStats) GetStoresBytesReadStat() map[uint64]uint64 {
+func (s *StoresStats) GetStoresBytesReadStat() map[uint64]float64 {
 	s.RLock()
 	defer s.RUnlock()
-	res := make(map[uint64]uint64, len(s.rollingStoresStats))
+	res := make(map[uint64]float64, len(s.rollingStoresStats))
 	for storeID, stats := range s.rollingStoresStats {
 		_, readRate := stats.GetBytesRate()
-		res[storeID] = uint64(readRate)
+		res[storeID] = readRate
 	}
 	return res
 }
 
 // GetStoresKeysWriteStat returns the keys write stat of all StoreInfo.
-func (s *StoresStats) GetStoresKeysWriteStat() map[uint64]uint64 {
+func (s *StoresStats) GetStoresKeysWriteStat() map[uint64]float64 {
 	s.RLock()
 	defer s.RUnlock()
-	res := make(map[uint64]uint64, len(s.rollingStoresStats))
+	res := make(map[uint64]float64, len(s.rollingStoresStats))
 	for storeID, stats := range s.rollingStoresStats {
-		res[storeID] = uint64(stats.GetKeysWriteRate())
+		res[storeID] = stats.GetKeysWriteRate()
 	}
 	return res
 }
 
 // GetStoresKeysReadStat returns the bytes read stat of all StoreInfo.
-func (s *StoresStats) GetStoresKeysReadStat() map[uint64]uint64 {
+func (s *StoresStats) GetStoresKeysReadStat() map[uint64]float64 {
 	s.RLock()
 	defer s.RUnlock()
-	res := make(map[uint64]uint64, len(s.rollingStoresStats))
+	res := make(map[uint64]float64, len(s.rollingStoresStats))
 	for storeID, stats := range s.rollingStoresStats {
-		res[storeID] = uint64(stats.GetKeysReadRate())
+		res[storeID] = stats.GetKeysReadRate()
 	}
 	return res
 }
